@@ -1,39 +1,26 @@
 {
-  "name": "realtime-editor",
+  "name": "client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@mui/material": "^5.14.4",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "axios": "^1.4.0",
-    "codemirror": "^5.65.14",
-    "express": "^4.17.3",
-    "react": "^17.0.2",
-    "react-avatar": "^4.0.0",
-    "react-dom": "^17.0.2",
-    "react-hot-toast": "^2.2.0",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
-    "react-select": "^5.7.4",
-    "socket.io": "^4.4.1",
-    "socket.io-client": "^4.4.1",
-    "uuid": "^8.3.2",
-    "web-vitals": "^2.1.4",
-    "y-codemirror": "^3.0.1",
-    "y-webrtc": "^10.2.5",
-    "y-websocket": "^1.5.0",
-    "yjs": "^13.6.7"
+    "bootstrap": "^5.3.1",
+    "codemirror": "^5.65.15",
+    "react": "^18.2.0",
+    "react-avatar": "^5.0.3",
+    "react-dom": "^18.2.0",
+    "react-hot-toast": "^2.4.1",
+    "react-router-dom": "^6.15.0",
+    "react-scripts": "5.0.1",
+    "socket.io-client": "^4.7.2",
+    "uuid": "^9.0.0",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start:front": "react-scripts start",
-    "start": "react-scripts --openssl-legacy-provider start && npm run server:prod",
+    "start": "react-scripts start",
     "build": "react-scripts build",
-    "server:dev": "nodemon server.js",
-    "server:prod": "node server.js",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
@@ -54,8 +41,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.15"
   }
 }
