@@ -1,10 +1,10 @@
-const ACTIONS = {
-    JOIN: 'join',
-    JOINED: 'joined',
-    DISCONNECTED: 'disconnected',
-    CODE_CHANGE: 'code-change',
-    SYNC_CODE: 'sync-code',
-    LEAVE: 'leave',
-};
+// All the events
 
-module.exports = ACTIONS;
+export const ACTIONS = {
+  JOIN: "join",
+  JOINED: "joined",
+  DISCONNECTED: "disconnected",
+  CODE_CHANGE: "conde-change",
+  SYNC_CODE: "sync-code",
+  LEAVE: "leave",
+};
