@@ -1,35 +1,21 @@
 import './App.css';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { Routes, Route } from "react-router-dom";
+import Home from './components/Home';
+import EditorPage from './components/EditorPage';
 import { Toaster } from 'react-hot-toast';
-import Home from './pages/Home';
-import EditorPage from './pages/EditorPage';
 
 function App() {
-    return (
-        <>
-            <div>
-                <Toaster
-                    position="top-right"
-                    toastOptions={{
-                        success: {
-                            theme: {
-                                primary: '#4aed88',
-                            },
-                        },
-                    }}
-                ></Toaster>
-            </div>
-            <BrowserRouter>
-                <Routes>
-                    <Route path="/" element={<Home />}></Route>
-                    <Route
-                        path="/editor/:roomId"
-                        element={<EditorPage />}
-                    ></Route>
-                </Routes>
-            </BrowserRouter>
-        </>
-    );
+  return (
+    <>
+    <div>
+      <Toaster  position='top-center'></Toaster>
+    </div>
+    <Routes>
+     <Route path='/' element={ <Home /> } />
+     <Route path='/editor/:roomId' element={ <EditorPage /> } />
+    </Routes>
+    </>
+  );
 }
 
 export default App;
