@@ -1,67 +1,66 @@
-import React, { useEffect, useRef, useState } from 'react';
-import Codemirror from 'codemirror';
-import 'codemirror/lib/codemirror.css';
-import 'codemirror/theme/dracula.css';
-import 'codemirror/mode/javascript/javascript';
-import 'codemirror/addon/edit/closetag';
-import 'codemirror/addon/edit/closebrackets';
-import ACTIONS from '../Actions';
-import { Button } from '@mui/material';
+import React, { useEffect, useRef } from "react";
+import "codemirror/mode/javascript/javascript";
+import "codemirror/theme/dracula.css";
+import "codemirror/addon/edit/closetag";
+import "codemirror/addon/edit/closebrackets";
+import "codemirror/lib/codemirror.css";
+import CodeMirror from "codemirror";
+import { ACTIONS } from "../Actions";
 
-const Editor = ({ socketRef, roomId, onCodeChange }) => {
-    const editorRef = useRef(null);
-
-    useEffect(() => {
-        async function init() {
-            editorRef.current = Codemirror.fromTextArea(
-                document.getElementById('realtimeEditor'),
-                {
-                    mode: { name: 'javascript', json: true },
-                    theme: 'dracula',
-                    autoCloseTags: true,
-                    autoCloseBrackets: true,
-                    lineNumbers: true,
-                }
-            );
-
-            editorRef.current.on('change', (instance, changes) => {
-                const { origin } = changes;
-                const code = instance.getValue();
-                onCodeChange(code);
-
-                // Store the current code in local storage
-                localStorage.setItem('editorContent', code);
-
-                if (origin !== 'setValue') {
-                    socketRef.current.emit(ACTIONS.CODE_CHANGE, {
-                        roomId,
-                        code,
-                    });
-                }
-            });
+function Editor({ socketRef, roomId, onCodeChange }) {
+  const editorRef = useRef(null);
+  useEffect(() => {
+    const init = async () => {
+      const editor = CodeMirror.fromTextArea(
+        document.getElementById("realtimeEditor"),
+        {
+          mode: { name: "javascript", json: true },
+          theme: "dracula",
+          autoCloseTags: true,
+          autoCloseBrackets: true,
+          lineNumbers: true,
         }
-        init();
-    }, []);
+      );
+      // for sync the code
+      editorRef.current = editor;
 
-    useEffect(() => {
-        if (socketRef.current) {
-            socketRef.current.on(ACTIONS.CODE_CHANGE, ({ code }) => {
-                if (code !== null) {
-                    editorRef.current.setValue(code);
-                }
-            });
+      editor.setSize(null, "100%");
+      editorRef.current.on("change", (instance, changes) => {
+        // console.log("changes", instance ,  changes );
+        const { origin } = changes;
+        const code = instance.getValue(); // code has value which we write
+        onCodeChange(code);
+        if (origin !== "setValue") {
+          socketRef.current.emit(ACTIONS.CODE_CHANGE, {
+            roomId,
+            code,
+          });
         }
+      });
+    };
 
-        return () => {
-            socketRef.current.off(ACTIONS.CODE_CHANGE);
-        };
-    }, [socketRef.current]);
+    init();
+  }, []);
 
-    return (
-        <div>
-            <textarea id="realtimeEditor"></textarea>
-        </div>
-    );
-};
+  // data receive from server
+  useEffect(() => {
+    if (socketRef.current) {
+      socketRef.current.on(ACTIONS.CODE_CHANGE, ({ code }) => {
+        if (code !== null) {
+          editorRef.current.setValue(code);
+        }
+      });
+    }
+    return () => {
+      socketRef.current.off(ACTIONS.CODE_CHANGE);
+    };
+  }, [socketRef.current]);
+
+  return (
+    <div style={{ height: "600px" }}>
+      <textarea id="realtimeEditor"></textarea>
+    </div>
+  );
+}
 
 export default Editor;
