@@ -1,11 +1,11 @@
-import { io } from 'socket.io-client';
-console.log(process.env.REACT_APP_BACKEND_URL);
-export const initSocket = async () => {
+import {io} from 'socket.io-client';
+
+export const initSocket = async () =>{
     const options = {
         'force new connection': true,
-        reconnectionAttempt: 'Infinity',
+        reconnectionAttempts : 'Infinity',
         timeout: 10000,
         transports: ['websocket'],
     };
     return io(process.env.REACT_APP_BACKEND_URL, options);
-};
+}
